UefiPayloadPkg: Add payload MM privileged components
In the payload MM design, bootloaders and payloads behave natively
in terms of ABI, CPU feature enablement, etc, and our entrypoint
performs all needed tasks to obtain control and return to the bootloader.

As we've hooked the MM IPL, the PE entrypoint of the MM core is called
following the payload MM design, rather than the X64 ABI.
MmCorePayloadMmEntryPoint handles the needed tasks for first-launch.

BlSmmCpuPayloadMm is an MM CPU driver, which means that it glues the
bootloader and payload together for runtime, among other responsibilities.

Signed-off-by: Benjamin Doron <benjamin.doron@9elements.com>