UefiPayloadPkg: Make "load" property reflect image, not whole FIT
Per the discussion in the UPL meeting on December 10, the load property
is for the FIT image, not whole FIT binary. To support Platform Init
implementations that load images individually, we fix this discrepancy.

However, EDK2-based implementations load the entire image at once.
Therefore, account for this when using the "load" property.

Signed-off-by: Benjamin Doron <benjamin.doron@9elements.com>